refactor(front): move inside specific component
