+++ packages/front/src/routes/auth/activate/[token]/components/AgentTypeStep/AgentTypeStep.controller.test.ts
