+++ packages/front/src/routes/auth/activate/[token]/components/AgentTypeStep/AgentTypeStep.controller.ts
