+++ packages/front/src/routes/auth/activate/[token]/components/AgentTypeStep/AgentTypeStep.svelte
