+++ packages/front/src/routes/auth/activate/[token]/components/AgentTypeStep/CollectedDataAlert.svelte
